Modify DTO/Deserializers/Proxies - Use Long instead of Integer for ID attributes

+++ src/main/java/com/achievementtracker/deserializer/AppListDeserializer.java
@@ -28,7 +28,7 @@ public class AppListDeserializer extends JsonDeserializer<AppListDTO> {
         for (JsonNode app : appsNode) {
             AppListDTO.AppDetailsDTO appDetailsDTO = new AppListDTO.AppDetailsDTO();
 
-            appDetailsDTO.setId(app.get("appid").intValue());
+            appDetailsDTO.setId(app.get("appid").longValue());
             appDetailsDTO.setName(app.get("name").textValue());
 
             apps.add(appDetailsDTO);
